requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.1
blinker==1.7.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
filelock==3.13.1
Flask==3.0.2
Flask-PyMongo==2.3.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.28.0
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
waitress==3.0.0
Werkzeug==3.0.1
yarl==1.9.4
